direct delivery changes

+++ Models/OpenPoModel.cs
@@ -1295,7 +1295,7 @@ namespace WMS.Models
 		public string mrnremarks { get; set; }
 		public int inwardid { get; set; }
 		public int acceptedqty { get; set; }
-		public int issuedqty { get; set; }
+		public decimal issuedqty { get; set; }
 
 	}
